remove unneeded ascription

+++ haskell/src/Euler/Problem/Problem10.hs
@@ -4,4 +4,4 @@ import Data.Numbers.Primes (primes)
 
 answer :: Integer
 answer = sum $ takeWhile (< 2*million) primes
-  where million = 10^(6::Integer)
+  where million = 10^6
